Cleaned up URLs and included Bldg ID in most URLs.
Previously, chart HTML and chart Data URLs were layered on top prior
URLs.  Now, they are added to the Base URL for the Application.
Also, some URLs used the string "one" when referring to requests that
referred to a one-building chart.  Now, the actual building ID is
substituted for the "one".  This is in preparation for moving to a set
of one-building charts that are available for every building, instead of
having to configure a unique set of charts for each building.