Added csv-export.js, highstock.js, HighCharts ver 4
The csv-export.js plug-in was added and allows for Excel and CSV export
direct from the Chart Menu.  The HighStock JS file was added to prepare
for its use.  Instead of hardcoding version 3.0.9, move back to
downloading the most recent HighCharts JS files (version 4.x), hoping
performance issue that occurred before was fixed.