Eliminated Caching of App JavaScript Files
Added a query parameter equal to the current time so a new file is
always requested.  Now, no need to rename the JS file every time a
change is made.  Lack of caching is not a problem, as page is rarely
reloaded and app JS files only take 200 ms or so to download.